+++ .env.example
@@ -1,5 +1,5 @@
 BASENAME=/
-BACKEND_URL=""
+VITE_BACKEND_URL=""
 
 DATABASE_URI="postgresql://gitpod:postgres@localhost:5432/example"
