+++ frontend/hooks/useGlobalReducer.jsx
@@ -1,33 +1,24 @@
+import { useContext, useReducer, createContext } from "react";
+import storeReducer, { initialStore } from '../store'; // Adjust the import path
 
+// Create a context to hold the global state of the application
+const StoreContext = createContext();
 
+// Define a provider component that encapsulates the store
+export function StoreProvider({ children }) {
+  // Initialize reducer with the initial state
+  const [store, dispatch] = useReducer(storeReducer, initialStore);
 
-// import React, { createContext, useReducer } from 'react';
-// import { storeReducer, initialState } from '../store'; // Adjust the import path
+  // Provide the store and dispatch method to all child components
+  return (
+    <StoreContext.Provider value={{ store, dispatch }}>
+      {children}
+    </StoreContext.Provider>
+  );
+}
 
-
-// import { useContext } from 'react';
-// import { StoreContext } from '../store'; // Adjust the import path
-
-// export default function useGlobalReducer() {
-//   const context = useContext(StoreContext);
-//   if (!context) {
-//     throw new Error('useGlobalReducer must be used within a StoreProvider');
-//   }
-//   return context;
-// }
-
-// // Create a context to hold the global state of the application
-// export const StoreContext = createContext();
-
-// // Define a provider component that encapsulates the store
-// export function StoreProvider({ children }) {
-//   // Initialize reducer with the initial state
-//   const [store, dispatch] = useReducer(storeReducer, initialState);
-
-//   // Provide the store and dispatch method to all child components
-//   return (
-//     <StoreContext.Provider value={{ store, dispatch }}>
-//       {children}
-//     </StoreContext.Provider>
-//   );
-// }
+// Custom hook to access the global state and dispatch function.
+export default function useGlobalReducer() {
+    const { dispatch, store } = useContext(StoreContext)
+    return { dispatch, store };
+}
