+++ frontend/main.jsx
@@ -5,7 +5,7 @@ import ReactDOM from 'react-dom/client';
 import './index.css'; // Global styles for your application
 import { RouterProvider } from 'react-router-dom'; // Import RouterProvider to use the router
 import { router } from './routes'; // Import the router configuration
-import { StoreProvider } from './store'; // Import the StoreProvider for global state management
+import { StoreProvider } from './hooks/useGlobalReducer'; // Import the StoreProvider for global state management
 import FetchInitialData from './FetchInitialData'; // Import FetchInitialData for initial data fetching
 import 'bootstrap/dist/css/bootstrap.min.css';
 import 'bootstrap/dist/js/bootstrap.bundle.min.js';
