+++ frontend/pages/Home.jsx
@@ -5,6 +5,7 @@ import { Link } from 'react-router-dom';
 import Login from "./Login.jsx";
 import "../style.css";
 import { Navigate, useNavigate } from "react-router-dom";
+import useGlobalReducer from '../hooks/useGlobalReducer';
