+++ frontend/routes.jsx
@@ -6,7 +6,7 @@ import {
     Route,
 } from "react-router-dom";
 import Layout from "./pages/Layout";
-import  {Home} from "./pages/Home";
+import Home from "./pages/Home";
 import FetchInitialData from "./FetchInitialData";
 import { MovieDetails } from "./components/MovieDetails";
 import PersonalQueue from "./pages/PersonalQueue";
